Created super class for both tables
