+++ src/database/ProposalsTable.py
@@ -16,50 +16,21 @@ from src.config.DatabaseConfig import site_table_name
 from pgvector.psycopg import register_vector
 
 from typing import List
+
+from src.database.ProposalsRequestsTable import ProposalsRequestsTable
 from src.database.data_types.ProposalRequest import ProposalRequest
 from src.database.data_types.ColumnNames import ColumnNames
 
 import psycopg
 
 
-class ProposalsTable:
+class ProposalsTable(ProposalsRequestsTable):
     __conn = DatabaseConnection()
 
     def __init__(self):
+        super().__init__()
         self.connection = self.__conn.get_instance().conn  # type: psycopg.connection.Connection
         self.all_string_columns_names = ColumnNames.all_proposal_string_columns_names
-
-    def add(self, proposal: ProposalRequest):
-        cursor = self.connection.cursor()
-        query  = proposal.get_insertion_query(site_table_name)
-        try:
-            cursor.execute(query)
-            cursor.close()
-        except Exception as e:
-            print(query)
-            print(e)
-        finally:
-            cursor.close()
-
-    def create_table_or_add_columns(self, table_name: str):
-        """
-        Creates table with name table_name if it does not exist with columns from ColumnNames.proposal_string_columns, id and embedding.
-        If table already exists, it adds columns from ColumnNames.proposal_string_columns and embedding if they do not exist.
-        Also registers vector extension.
-        """
-        self.connection.execute('CREATE EXTENSION IF NOT EXISTS vector')
-        proposal_string_columns = [
-            f"{column_name} {ColumnNames.types[column_name]}" for column_name in self.all_string_columns_names
-        ]
-        self.connection.execute(f"""
-            CREATE TABLE IF NOT EXISTS {table_name} 
-            (id SERIAL PRIMARY KEY, {", ".join(proposal_string_columns)}, embedding vector(1536))
-        """)
-        self.connection.execute(f'ALTER TABLE {table_name} ADD COLUMN IF NOT EXISTS embedding vector(1536)')
-        for proposal_string_column in proposal_string_columns:
-            self.connection.execute(f'ALTER TABLE {table_name} ADD COLUMN IF NOT EXISTS {proposal_string_column}')
-
-        register_vector(self.connection)
 
     def get_similar(self, request: ProposalRequest, start: int = 0, amount: int = 5) -> List[ProposalRequest]:
         embedding = request.embedding
